Bardo.ExperimentManager.ExperimentManagerClient (Bardo v0.1.0)

Client module for interacting with the ExperimentManager.

Functions

configure(experiment_id, config)

@spec configure(Bardo.Models.experiment_id(), map()) :: :ok | {:error, term()}

Configure an existing experiment.

get_best_solution(experiment_id)

@spec get_best_solution(Bardo.Models.experiment_id()) ::
  {:ok, term()} | {:error, term()}

Get the best solution from an experiment.

new_experiment(name)

@spec new_experiment(String.t()) ::
  {:ok, Bardo.Models.experiment_id()} | {:error, term()}

Create a new experiment.

run_complete(population_id, trace)

@spec run_complete(Bardo.Models.population_id(), Bardo.Models.trace()) :: :ok

Notifies the experiment manager that a run has completed with the given results.

start(experiment_id)

@spec start(Bardo.Models.experiment_id()) :: :ok | {:error, term()}

Start the experiment with the given ID.

For compatibility with older code.

start_evaluation(experiment_id, fitness_function)

@spec start_evaluation(Bardo.Models.experiment_id(), function()) ::
  :ok | {:error, term()}

Start evaluation with a fitness function.

start_run()

@spec start_run() :: :ok

Sends a message to start a new experiment run.

status(experiment_id)

@spec status(Bardo.Models.experiment_id()) ::
  {:in_progress, map()} | {:completed, map()} | {:error, term()}

Get the status of an experiment.

stop(experiment_id)

@spec stop(Bardo.Models.experiment_id()) :: :ok | {:error, term()}

Stop an experiment.