Bardo.PopulationManager.PopulationManagerClient (Bardo v0.1.0)

Client module for interacting with the PopulationManager.

Functions

agent_terminated(agent_id)

@spec agent_terminated(Bardo.Models.agent_id()) :: :ok

Notifies the population manager that an agent has terminated.

new_run()

@spec new_run() :: :ok

Sends a message to start a new population manager run.

restart_run()

@spec restart_run() :: :ok

Sends a message to restart a population manager run.

set_evaluations(specie_id, aea, cycle_acc, time_acc)

@spec set_evaluations(Bardo.Models.specie_id(), integer(), integer(), integer()) ::
  :ok

Sends evaluation data to the population manager.

set_goal_reached()

@spec set_goal_reached() :: :ok

Notifies the population manager that a goal has been reached.

set_op_tag(op_tag)

@spec set_op_tag(:pause | :continue | :stop) :: :ok

Sets the operation tag for the population manager.

stop(population_id)

@spec stop(atom() | binary()) :: :ok | {:error, :not_found}

Stops a population manager run with the specified ID.

Parameters

population_id - ID of the population to stop

Returns

:ok - If the population was stopped successfully

{:error, :not_found} - If the population was not found

validation_complete(agent_id, fitness)

@spec validation_complete(Bardo.Models.agent_id(), float()) :: :ok

Notifies the population manager that validation is complete with the given fitness.